aws_tools/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==2.5.2               # via kombu
billiard==3.6.1.0         # via celery
boto3==1.10.15
botocore==1.13.15         # via boto3, s3transfer
celery==4.3.0
certifi==2019.9.11        # via requests
cffi==1.13.2              # via cryptography
chardet==3.0.4            # via requests
cryptography==2.8         # via josepy, mozilla-django-oidc, pyopenssl
django-celery-beat==1.5.0
django-choices==1.7.0
django-filter==2.2.0
django-netfields==1.1.1
django-timezone-field==3.1  # via django-celery-beat
django==2.2.7
djangorestframework==3.10.3
docutils==0.15.2          # via botocore
idna==2.8                 # via requests
importlib-metadata==0.23  # via kombu
jmespath==0.9.4           # via boto3, botocore
josepy==1.2.0             # via mozilla-django-oidc
kombu==4.6.6              # via celery
more-itertools==7.2.0     # via zipp
mozilla-django-oidc==1.2.2
netaddr==0.7.19           # via django-netfields
pycparser==2.19           # via cffi
pyopenssl==19.0.0         # via josepy
python-crontab==2.4.0     # via django-celery-beat
python-dateutil==2.8.0    # via botocore, python-crontab
pytz==2019.3              # via celery, django, django-timezone-field
requests==2.22.0          # via mozilla-django-oidc
s3transfer==0.2.1         # via boto3
six==1.13.0               # via cryptography, josepy, pyopenssl, python-dateutil
sqlparse==0.3.0           # via django
urllib3==1.25.7           # via botocore, requests
vine==1.3.0               # via amqp, celery
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.6.0        # via josepy
